fix cycle-tool (and add it to skeleton-only)
