+++ frontend/javascripts/oxalis/model/reducers/annotation_reducer.js
@@ -11,6 +11,7 @@ import {
 } from "oxalis/model/reducers/reducer_helpers";
 import { AnnotationToolEnum } from "oxalis/constants";
 import { hideBrushReducer } from "oxalis/model/reducers/volumetracing_reducer_helpers";
+import { getDisabledInfoForTools } from "oxalis/model/accessors/tool_accessor";
 
 const updateTracing = (state: OxalisState, shape: StateShape1<"tracing">): OxalisState =>
   updateKey(state, "tracing", shape);
@@ -156,11 +157,26 @@ function AnnotationReducer(state: OxalisState, action: Action): OxalisState {
       if (!state.tracing.restrictions.allowUpdate) {
         return state;
       }
+
+      const disabledToolInfo = getDisabledInfoForTools(state);
+
       const tools = Object.keys(AnnotationToolEnum);
       const currentToolIndex = tools.indexOf(state.tracing.activeTool);
-      const newTool = tools[(currentToolIndex + 1) % tools.length];
+      let newTool;
 
-      return setToolReducer(hideBrushReducer(state), newTool);
+      // Search for the next tool which is not disabled.
+      for (
+        let newToolIndex = currentToolIndex + 1;
+        newToolIndex < currentToolIndex + tools.length;
+        newToolIndex++
+      ) {
+        newTool = tools[newToolIndex % tools.length];
+        if (!disabledToolInfo[newTool].isDisabled) {
+          return setToolReducer(hideBrushReducer(state), newTool);
+        }
+      }
+      // Don't change the current tool if another tool could not be selected.
+      return state;
     }
 
     default:
